feat: cria componente com par de sliders da tela my-profile (Ref. #283)
